refactor: local sample file helpers
