+++ scripts/dev_seed_and_preview.py
@@ -18,33 +18,8 @@ from app.processors.visit_html_parser import extract_visit_summaries
 from app.processors.structuring import insert_lab_results, insert_visit_summaries
 from app.prompts.summarizer import summarize_blocks
 from app.api.rag import ask_question, QueryRequest
+from app.utils.sample_files import create_sample_pdf, create_sample_html
 
-
-def create_sample_pdf(path: Path) -> None:
-    """Create a small PDF with two lab lines for demo purposes."""
-    import fitz  # PyMuPDF
-
-    doc = fitz.open()
-    page = doc.new_page()
-    page.insert_text((72, 72), "Cholesterol 5.8 mmol/L 2023-05-01")
-    page.insert_text((72, 90), "Hemoglobin 13.5 g/dL 2023-05-02")
-    doc.save(path)
-    doc.close()
-
-
-def create_sample_html(path: Path) -> None:
-    """Create simple HTML with a single visit summary."""
-    html = """
-    <html><body>
-      <div class='visit'>
-        <span class='date'>2023-06-01</span>
-        <span class='provider'>General Hospital</span>
-        <span class='doctor'>Dr. Jones</span>
-        <p class='notes'>Routine check</p>
-      </div>
-    </body></html>
-    """
-    path.write_text(html, encoding="utf-8")
 
 
 def _maybe_mock_openai() -> None:
